requirements.txt
Pillow
pymediainfo
flask
flask_restful
TgCrypto
heroku3
aiohttp
requests
random2
asyncio
gtts
googlesearch-python
google
beautifulsoup4
google-api-python-client
lyricsgenius
telegraph
wget
youtube-search-python
yt-dlp==2024.03.10
pyrogram==2.0.105
openai==0.28
pytgcalls==0.8.6
